WeeeeoooooHooooooooooooooooooo!
Have mapping working off gps coords, and I must say it's looking hot.  Had to change projection of the maps to init=nad83:2001 and units meters, but that did the trick.  Map now opens to MIT, and should be able to take gps incoming to portray location.  We are now set to do waypoint following.